+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 
+- Fix compatibility with Matplotlib 3.6.
 - Non-multiple cursors can now be dragged.
 
 0.5.1